+++ apps/demo-worker/src/app.module.ts
@@ -1,7 +1,5 @@
 import { Module } from '@nestjs/common';
 import { BullModule } from '@nestjs/bull';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { TranscribeModule } from './transcribe/transcribe.module';
 
 @Module({
@@ -12,12 +10,7 @@ import { TranscribeModule } from './transcribe/transcribe.module';
         port: 6379,
       },
     }),
-    BullModule.registerQueue({
-      name: 'todo',
-    }),
     TranscribeModule,
   ],
-  controllers: [AppController],
-  providers: [AppService],
 })
 export class AppModule {}
