--- apps/demo-worker/src/app.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class AppService {
-  getStatus(): { status: boolean } {
-    return { status: true };
-  }
-}
