+++ apps/web/app/(routes)/(auth)/sign-in/page.tsx
@@ -1,0 +1,10 @@
+import { SignIn } from '@clerk/nextjs';
+
+const SignInPage = () => {
+  return (
+    <div className="flex w-full justify-center">
+      <SignIn path="/sign-in" routing="path" signUpUrl="/sign-up" />
+    </div>
+  );
+};
+export default SignInPage;
