+++ apps/web/app/(routes)/(auth)/user-profile/page.tsx
@@ -1,0 +1,9 @@
+import { UserProfile } from '@clerk/nextjs';
+
+const UserProfilePage = () => (
+  <div className="flex w-full justify-center">
+    <UserProfile path="/user-profile" routing="path" />
+  </div>
+);
+
+export default UserProfilePage;
