+++ apps/web/codegen.yml
@@ -1,7 +1,7 @@
 overwrite: true
-schema: 'http://localhost:3000/graphql' # TODO Replace with your GraphQL API endpoint
+schema: 'http://localhost:5000/graphql'
 generates:
-  ./src/graphql/generated/schema.ts:
+  ./app/_graphql/generated/schema.ts:
     plugins:
       - 'typescript'
       - 'typescript-resolvers'
