+++ apps/web/package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
+    "predev": "npx graphql-codegen",
     "dev": "next dev --port 3000",
     "build": "next build",
     "start": "next start",
@@ -12,37 +13,29 @@
     "lint:next": "next lint",
     "lint:typescript": "tsc --noEmit",
     "lint:prettier": "prettier --check ./src",
-    "format": "npm run && npm run format:prettier",
+    "format": "npm run format:prettier",
     "format:prettier": "prettier --write ./src"
   },
   "dependencies": {
     "@apollo/client": "^3.7.12",
     "@clerk/nextjs": "^4.16.4",
-    "@package/header": "*",
-    "@tanstack/react-query": "^4.3.8",
-    "@trpc/client": "^10.7.0",
-    "@trpc/next": "^10.7.0",
-    "@trpc/react-query": "^10.7.0",
-    "@trpc/server": "^10.7.0",
-    "graphql": "^16.6.0",
-    "graphql-scalars": "^1.20.1",
-    "next": "^13.0.0",
+    "@graphql-codegen/typescript-resolvers": "^4.0.1",
+    "@package/nextjs-header": "*",
+    "graphql": "^16.7.1",
+    "next": "^13.4.12",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "superjson": "^1.7.4",
-    "zod": "^3.20.2"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^3.0.0",
-    "@graphql-codegen/client-preset": "^2.0.0",
+    "@graphql-codegen/cli": "^4.0.1",
+    "@graphql-codegen/client-preset": "^4.1.0",
     "@tooling/eslint-config-custom": "*",
     "@tooling/jest": "*",
     "@tooling/typescript": "*",
     "@types/node": "^18.7.20",
     "@types/react": "^18.0.9",
     "autoprefixer": "^10.4.14",
-    "tailwindcss": "^3.2.4",
-    "tsx": "^3.9.0",
+    "tailwindcss": "^3.3.3",
     "typescript": "^4.9.5"
   },
   "engines": {
