--- apps/web/src/pages/sign-in/index.tsx
@@ -1,4 +1,0 @@
-import { SignIn } from '@clerk/nextjs';
-
-const SignInPage = () => <SignIn path="/sign-in" routing="path" signUpUrl="/sign-up" />;
-export default SignInPage;
