--- apps/web/src/pages/sign-up/index.tsx
@@ -1,4 +1,0 @@
-import { SignUp } from '@clerk/nextjs';
-
-const SignUpPage = () => <SignUp path="/sign-up" routing="path" signInUrl="/sign-in" />;
-export default SignUpPage;
