--- apps/web/src/pages/user-profile/index.tsx
@@ -1,5 +1,0 @@
-import { UserProfile } from '@clerk/nextjs';
-
-const UserProfilePage = () => <UserProfile path="/user-profile" routing="path" />;
-
-export default UserProfilePage;
