--- apps/web/src/test/example/sum.ts
@@ -1,1 +1,0 @@
-export const sum = (...a: number[]) => a.reduce((acc, val) => acc + val, 0);
