--- packages/auth/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@package/auth",
-  "version": "1.0.0",
-  "main": "src/index.ts",
-  "private": true,
-  "dependencies": {
-    "@nestjs/graphql": "^11.0.6",
-    "@nestjs/passport": "^9.0.3",
-    "eslint-config-custom": "*",
-    "jwks-rsa": "^3.0.1",
-    "passport-jwt": "^4.0.1",
-    "tsconfig": "*"
-  },
-  "publishConfig": {
-    "access": "restricted"
-  }
-}
