--- packages/auth/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./jwt-auth.guard";
-export * from "./jwt.strategy";
-export * from "./current-user.decorator";
