--- packages/auth/src/jwt-auth.guard.ts
@@ -1,25 +1,0 @@
-import { ExecutionContext, Injectable } from "@nestjs/common";
-import { GqlExecutionContext } from "@nestjs/graphql";
-import { Reflector } from "@nestjs/core";
-import { AuthGuard } from "@nestjs/passport";
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard("jwt") {
-  constructor(private reflector: Reflector) {
-    super();
-    console.log(reflector);
-  }
-
-  canActivate(context: ExecutionContext) {
-    const allowUnauthorizedRequest = this.reflector.get<boolean>(
-      "allowUnauthorizedRequest",
-      context.getHandler()
-    );
-    return allowUnauthorizedRequest || super.canActivate(context);
-  }
-
-  getRequest(context: ExecutionContext) {
-    const ctx = GqlExecutionContext.create(context);
-    return ctx.getContext().req;
-  }
-}
