--- packages/fvst-logo/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@package/fvst-logo",
-  "version": "10.7.0",
-  "private": true,
-  "dependencies": {
-    "eslint-config-custom": "*",
-    "tsconfig": "*"
-  },
-  "publishConfig": {
-    "access": "restricted"
-  }
-}
