--- packages/fvst-logo/src/index.tsx
@@ -1,37 +1,0 @@
-import * as React from "react";
-import { SVGProps } from "react";
-
-export const Logo = (props: SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    xmlnsXlink="http://www.w3.org/1999/xlink"
-    width="130px"
-    height="40px"
-    viewBox="0 0 100 31"
-    {...props}
-  >
-    <title>FVST</title>
-    <g id="surface1">
-      <path
-        className="fill-fvst-orange"
-        d="M 0 0.34375 L 20.34375 0.34375 L 20.34375 5.476562 L 6.066406 5.476562 L 6.066406 30.589844 L 0 30.589844 Z M 9.9375 18.464844 L 7.140625 12.851562 L 16.691406 12.851562 L 16.691406 18.464844 Z M 9.9375 18.464844 "
-      />
-      <path
-        className="fill-fvst-orange"
-        d="M 65.429688 12.945312 C 66.804688 13.320312 68.070312 13.742188 69.238281 14.191406 C 70.398438 14.652344 71.398438 15.21875 72.246094 15.878906 C 73.089844 16.535156 73.746094 17.332031 74.226562 18.253906 C 74.699219 19.175781 74.9375 20.265625 74.9375 21.53125 L 74.9375 21.65625 C 74.9375 23.039062 74.65625 24.308594 74.101562 25.460938 C 73.539062 26.613281 72.765625 27.597656 71.785156 28.414062 C 70.792969 29.230469 69.632812 29.867188 68.296875 30.308594 C 66.960938 30.757812 65.511719 30.980469 63.933594 30.980469 C 62.472656 30.980469 61.074219 30.800781 59.738281 30.445312 C 58.402344 30.089844 57.191406 29.617188 56.105469 29.03125 C 55.019531 28.445312 54.058594 27.785156 53.246094 27.042969 C 52.433594 26.296875 51.804688 25.523438 51.378906 24.726562 L 56.410156 21.53125 C 56.953125 22.140625 57.496094 22.675781 58.046875 23.15625 C 58.589844 23.628906 59.164062 24.035156 59.769531 24.359375 C 60.375 24.695312 61.035156 24.945312 61.753906 25.113281 C 62.472656 25.292969 63.257812 25.375 64.125 25.375 C 64.664062 25.375 65.21875 25.324219 65.78125 25.207031 C 66.347656 25.09375 66.847656 24.90625 67.308594 24.644531 C 67.765625 24.382812 68.140625 24.054688 68.421875 23.667969 C 68.707031 23.28125 68.851562 22.800781 68.851562 22.222656 C 68.851562 21.816406 68.746094 21.425781 68.550781 21.0625 C 68.351562 20.683594 68.007812 20.316406 67.515625 19.960938 C 67.023438 19.605469 66.367188 19.261719 65.53125 18.925781 C 64.699219 18.589844 63.652344 18.265625 62.390625 17.949219 C 60.84375 17.574219 59.445312 17.15625 58.214844 16.671875 C 56.984375 16.203125 55.941406 15.625 55.074219 14.945312 C 54.21875 14.265625 53.550781 13.457031 53.089844 12.503906 C 52.628906 11.550781 52.402344 10.429688 52.402344 9.132812 C 52.402344 7.71875 52.660156 6.441406 53.195312 5.308594 C 53.726562 4.167969 54.46875 3.214844 55.40625 2.441406 C 56.355469 1.664062 57.464844 1.066406 58.738281 0.648438 C 60.011719 0.230469 61.378906 0.0195312 62.839844 0.0195312 C 64.363281 0.0195312 65.730469 0.1875 66.972656 0.535156 C 68.203125 0.878906 69.3125 1.320312 70.28125 1.855469 C 71.253906 2.386719 72.109375 2.996094 72.839844 3.664062 C 73.570312 4.347656 74.175781 4.996094 74.664062 5.625 L 69.730469 8.808594 C 68.496094 7.625 67.328125 6.796875 66.222656 6.324219 C 65.113281 5.855469 63.964844 5.613281 62.753906 5.613281 C 62.015625 5.613281 61.367188 5.707031 60.84375 5.894531 C 60.3125 6.085938 59.875 6.316406 59.53125 6.585938 C 59.1875 6.859375 58.925781 7.175781 58.738281 7.519531 C 58.550781 7.863281 58.457031 8.199219 58.457031 8.515625 C 58.457031 9.089844 58.601562 9.582031 58.902344 10 C 59.207031 10.421875 59.644531 10.796875 60.238281 11.144531 C 60.824219 11.488281 61.554688 11.804688 62.433594 12.097656 C 63.320312 12.367188 64.3125 12.660156 65.429688 12.945312 Z M 65.429688 12.945312 "
-      />
-      <path
-        className="fill-fvst-orange"
-        d="M 91.136719 0.386719 L 91.136719 30.589844 L 84.949219 30.589844 L 84.949219 5.65625 L 76.085938 5.65625 L 76.085938 0.386719 Z M 92.046875 0.386719 L 100 0.386719 L 100 5.65625 L 94.664062 5.65625 Z M 92.046875 0.386719 "
-      />
-      <path
-        className="fill-fvst-orange"
-        d="M 27.503906 0.386719 L 21.054688 0.386719 L 31.335938 25.679688 L 36.199219 21.492188 Z M 27.503906 0.386719 "
-      />
-      <path
-        className="fill-fvst-orange"
-        d="M 52.234375 0.386719 L 39.894531 30.589844 L 33.402344 30.589844 L 41.355469 11.132812 L 36.648438 11.132812 L 33.957031 5.738281 L 43.507812 5.738281 L 45.78125 0.386719 Z M 52.234375 0.386719 "
-      />
-    </g>
-  </svg>
-);
