--- packages/header/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@package/header",
-  "version": "10.7.0",
-  "private": true,
-  "dependencies": {
-    "@clerk/clerk-react": "^4.15.3",
-    "@clerk/nextjs": "^4.16.4",
-    "@headlessui/react": "^1.7.8",
-    "@heroicons/react": "^2.0.14",
-    "@package/fvst-logo": "*",
-    "@tooling/typescript": "*",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
-    "eslint-config-custom": "*",
-    "react": "^18.2.0",
-    "tailwindcss": "^3.2.4",
-    "tsconfig": "*"
-  },
-  "publishConfig": {
-    "access": "restricted"
-  }
-}
