--- packages/header/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@tooling/typescript/react-library.json",
-  "include": ["."],
-  "exclude": ["dist", "build", "node_modules"]
-}
