+++ packages/nestjs-health/.prettierrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@tooling/prettier/.prettierrc');
