+++ packages/nextjs-header/.gitignore
