+++ packages/nextjs-header/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@package/nextjs-header",
+  "version": "10.7.0",
+  "main": "src/index.tsx",
+  "private": true,
+  "dependencies": {
+    "@clerk/clerk-react": "^4.15.3",
+    "@clerk/nextjs": "^4.16.4",
+    "@headlessui/react": "^1.7.8",
+    "@heroicons/react": "^2.0.14",
+    "react": "^18.2.0",
+    "tailwindcss": "^3.2.4"
+  },
+  "devDependencies": {
+    "@tooling/typescript": "*",
+    "@tooling/prettier": "*",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9"
+  },
+  "publishConfig": {
+    "access": "restricted"
+  }
+}
