+++ packages/nextjs-header/src/components/drop-down-menu.tsx
@@ -1,0 +1,75 @@
+import React from 'react';
+
+import { Popover, Transition } from '@headlessui/react';
+import { ChevronDownIcon } from '@heroicons/react/20/solid';
+
+function classNames(...classes: string[]) {
+  return classes.filter(Boolean).join(' ');
+}
+
+type Props = {
+  title: string;
+  items: {
+    name: string;
+    description: string;
+    href: string;
+    icon: ({ className }: { className: string }) => React.ReactElement | null;
+  }[];
+};
+
+export const DropDownMenu: React.FC<Props> = ({ title, items }) => {
+  return (
+    <Popover.Group as="nav" className="relative space-x-10 md:flex">
+      <Popover className="relative">
+        {({ open }) => (
+          <>
+            <Popover.Button
+              className={classNames(
+                open ? 'text-fvst-grey' : 'text-white',
+                'hover:text-fvst-grey group inline-flex items-center border-0 text-base font-light'
+              )}
+            >
+              <span>{title}</span>
+              <ChevronDownIcon
+                className={classNames(
+                  open ? 'text-gray-600' : 'text-gray-400',
+                  'ml-2 h-5 w-5 group-hover:text-gray-500'
+                )}
+                aria-hidden="true"
+              />
+            </Popover.Button>
+            <Transition
+              as={React.Fragment}
+              enter="transition ease-out duration-200"
+              enterFrom="opacity-0 translate-y-1"
+              enterTo="opacity-100 translate-y-0"
+              leave="transition ease-in duration-150"
+              leaveFrom="opacity-100 translate-y-0"
+              leaveTo="opacity-0 translate-y-1"
+            >
+              <Popover.Panel className="absolute z-10 -ml-4 mt-3 w-screen max-w-md transform px-2 sm:px-0 lg:left-1/2 lg:ml-0 lg:-translate-x-1/2">
+                <div className="overflow-hidden shadow-lg ring-1 ring-black ring-opacity-5">
+                  <div className="relative grid gap-6 bg-white px-5 py-6 sm:gap-8 sm:p-8">
+                    {items.map((item) => (
+                      <a
+                        key={item.name}
+                        href={item.href}
+                        className="-m-3 flex items-start rounded-lg p-3 hover:bg-gray-50"
+                      >
+                        <item.icon className="h-6 w-6 flex-shrink-0 text-indigo-600" />
+                        <div className="ml-4">
+                          <p className="text-base font-medium text-gray-900">{item.name}</p>
+                          <p className="mt-1 text-sm text-gray-500">{item.description}</p>
+                        </div>
+                      </a>
+                    ))}
+                  </div>
+                </div>
+              </Popover.Panel>
+            </Transition>
+          </>
+        )}
+      </Popover>
+    </Popover.Group>
+  );
+};
