+++ packages/nextjs-header/src/components/user-button.tsx
@@ -1,0 +1,24 @@
+import { SignedIn, SignedOut, SignInButton, UserButton as ClerkUserButton } from '@clerk/clerk-react';
+
+export const UserButton = () => {
+  return (
+    <>
+      <SignedIn>
+        {/* Mount the UserButton component */}
+        <ClerkUserButton
+          appearance={{
+            userProfile: {
+              variables: {
+                borderRadius: '0px',
+              },
+            },
+          }}
+        />
+      </SignedIn>
+      <SignedOut>
+        {/* Signed-out users get sign in button */}
+        <SignInButton />
+      </SignedOut>
+    </>
+  );
+};
