+++ packages/nextjs-header/src/index.tsx
@@ -1,0 +1,1 @@
+export * from './Header';
